requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
boto3==1.28.62
botocore==1.31.62
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cmaes==0.10.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.1.1
CouchDB==1.2
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.6.1
decorator==5.1.1
docopt==0.6.2
einops==0.7.0
ffmpeg-python==0.2.0
filelock==3.12.4
firebase-admin==6.2.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
huggingface-hub==0.17.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
julius==0.2.7
kiwisolver==1.4.5
langchain==0.0.313
langsmith==0.0.43
lazy_loader==0.3
librosa==0.10.1
lightning==2.1.0
lightning-utilities==0.9.0
llvmlite==0.41.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
omegaconf==2.3.0
onnxruntime==1.16.1
openai==0.28.1
openai-whisper==20230918
optuna==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
primePy==1.3
proto-plus==1.22.3
protobuf==4.24.4
pyannote.audio==3.0.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==2.1.0
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
semver==3.0.2
sentencepiece==0.1.99
shellingham==1.5.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.15
SQLAlchemy==2.0.21
stable-ts==2.12.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tiktoken==0.3.3
tokenizers==0.14.1
torch==2.1.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.1.0
torchmetrics==1.2.0
tqdm==4.66.1
transformers==4.34.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.0
yarl==1.9.2
